+++ CHANGELOG.rst
@@ -16,7 +16,7 @@ Added
     - ``django-axes`` support for login security (#1976, #2334)
     - ``app_landingzones_transfer.rst`` landing zone file deletion note (#2478)
 - **Samplesheets**
-    - Vue3 app (#994, #2389, #2390, #2403, #2457, #2458, #2459)
+    - Vue3 app (#994, #2389, #2390, #2403, #2431, #2457, #2458, #2459)
     - Vite integration using ``django-vite`` (#994)
     - ``use_vue3_app`` app setting (#994)
     - ``fixirodstickets`` management command (#2461)