+++ CHANGELOG.rst
@@ -25,8 +25,8 @@ Changed
 -------
 
 - **General**
-    - Upgrade to Django v5.2.13 (#2317, #2334)
-    - Upgrade to django-sodar-core v1.4.2 (#2317, #2334, #2358, #2375)
+    - Upgrade to Django v5.2.17 (#2317, #2334)
+    - Upgrade to django-sodar-core v1.4.3 (#2317, #2334, #2358, #2375, #2505)
     - Upgrade to python-irodsclient v3.3.0 (#2253, #2471)
     - Upgrade to django-rest-knox v5.0 (#2334)
     - Upgrade to tooz v9.0 (#2485)